requirements.txt
git+https://github.com/TEEN-BOOM/fractal.git
discord.py
pandas
Pillow
psycopg2
praw
PyNaCl
aiohttp
async-timeout
attrs
cffi
chardet
dnspython
idna
multidict
numpy
pycparser
python-dateutil
pytz
six
websockets
requests
youtube_dl